+++ src/lib/components/element/ThemeSwitch.svelte
@@ -4,7 +4,7 @@
   // onMount(() => {
   //   document.querySelector('html')?.classList.add('dark');
   // });
-  import Icon from '$lib/Icon.svelte';
+  import Icon from '$lib/components/element/Icon.svelte';
   import { faMoon, faSun } from '@fortawesome/free-solid-svg-icons';
   import { onMount } from 'svelte';
