+++ src/lib/components/modules/SkillBox.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import SimpleIcon from '$lib/components/element/SimpleIcon.svelte';
-  import Icon from '$lib/Icon.svelte';
+  import Icon from '$lib/components/element/Icon.svelte';
   import { faEllipsis } from '@fortawesome/free-solid-svg-icons';
 
   export let name = '';
