+++ src/routes/+layout.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import '../app.postcss';
   import { onMount } from 'svelte';
-  import Header from '$lib/header/Header.svelte';
+  import Header from '$lib/templates/Header.svelte';
   import Loading from '$lib/components/element/Loading.svelte';
   import Footer from '$lib/templates/Footer.svelte';
   export let onLoad = false;
